sync from the upstream

+++ ql/src/java/org/apache/hadoop/hive/ql/session/OperationLog.java
@@ -1,4 +1,4 @@
-/**
+/*
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
  * distributed with this work for additional information
@@ -18,10 +18,10 @@
 package org.apache.hadoop.hive.ql.session;
 
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
 import org.apache.hadoop.hive.conf.HiveConf;
 import org.apache.hadoop.io.IOUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.*;
 import java.sql.SQLException;
@@ -33,23 +33,48 @@ import java.util.List;
  * for accessing, reading, writing, and removing the file.
  */
 public class OperationLog {
-  private static final Log LOG = LogFactory.getLog(OperationLog.class.getName());
+  private static final Logger LOG = LoggerFactory.getLogger(OperationLog.class.getName());
 
   private final String operationName;
   private final LogFile logFile;
+  // If in test mode then the LogDivertAppenderForTest created an extra log file containing only
+  // the output needed for the qfile results.
+  private final LogFile testLogFile;
+  // True if we are running test and the extra test file should be used when the logs are
+  // requested.
+  private final boolean isShortLogs;
+  // True if the logs should be removed after the operation. Should be used only in test mode
+  private final boolean isRemoveLogs;
   private LoggingLevel opLoggingLevel = LoggingLevel.UNKNOWN;
 
-  public static enum LoggingLevel {
+  public enum LoggingLevel {
     NONE, EXECUTION, PERFORMANCE, VERBOSE, UNKNOWN
   }
 
-  public OperationLog(String name, File file, HiveConf hiveConf) throws FileNotFoundException {
+  public OperationLog(String name, File file, HiveConf hiveConf) {
     operationName = name;
     logFile = new LogFile(file);
 
     if (hiveConf.getBoolVar(HiveConf.ConfVars.HIVE_SERVER2_LOGGING_OPERATION_ENABLED)) {
       String logLevel = hiveConf.getVar(HiveConf.ConfVars.HIVE_SERVER2_LOGGING_OPERATION_LEVEL);
       opLoggingLevel = getLoggingLevel(logLevel);
+    }
+
+    // If in test mod create a test log file which will contain only logs which are supposed to
+    // be written to the qtest output
+    if (hiveConf.getBoolVar(HiveConf.ConfVars.HIVE_IN_TEST)) {
+      isRemoveLogs = hiveConf.getBoolVar(HiveConf.ConfVars.HIVE_TESTING_REMOVE_LOGS);
+      if (hiveConf.getBoolVar(HiveConf.ConfVars.HIVE_TESTING_SHORT_LOGS)) {
+        testLogFile = new LogFile(new File(file.getAbsolutePath() + ".test"));
+        isShortLogs = true;
+      } else {
+        testLogFile = null;
+        isShortLogs = false;
+      }
+    } else {
+      testLogFile = null;
+      isShortLogs = false;
+      isRemoveLogs = true;
     }
   }
 
@@ -72,37 +97,6 @@ public class OperationLog {
   }
 
   /**
-   * Singleton OperationLog object per thread.
-   */
-  private static final ThreadLocal<OperationLog> THREAD_LOCAL_OPERATION_LOG = new
-      ThreadLocal<OperationLog>() {
-    @Override
-    protected OperationLog initialValue() {
-      return null;
-    }
-  };
-
-  public static void setCurrentOperationLog(OperationLog operationLog) {
-    THREAD_LOCAL_OPERATION_LOG.set(operationLog);
-  }
-
-  public static OperationLog getCurrentOperationLog() {
-    return THREAD_LOCAL_OPERATION_LOG.get();
-  }
-
-  public static void removeCurrentOperationLog() {
-    THREAD_LOCAL_OPERATION_LOG.remove();
-  }
-
-  /**
-   * Write operation execution logs into log file
-   * @param operationLogMessage one line of log emitted from log4j
-   */
-  public void writeOperationLog(String operationLogMessage) {
-    logFile.write(operationLogMessage);
-  }
-
-  /**
    * Read operation execution logs from log file
    * @param isFetchFirst true if the Enum FetchOrientation value is Fetch_First
    * @param maxRows the max number of fetched lines from log
@@ -110,36 +104,38 @@ public class OperationLog {
    * @throws java.sql.SQLException
    */
   public List<String> readOperationLog(boolean isFetchFirst, long maxRows)
-      throws SQLException{
-    return logFile.read(isFetchFirst, maxRows);
+      throws SQLException {
+    if (isShortLogs) {
+      return testLogFile.read(isFetchFirst, maxRows);
+    } else {
+      return logFile.read(isFetchFirst, maxRows);
+    }
   }
 
   /**
    * Close this OperationLog when operation is closed. The log file will be removed.
    */
   public void close() {
-    logFile.remove();
+    if (isShortLogs) {
+      // In case of test, do just close the log files, do not remove them.
+      logFile.close(isRemoveLogs);
+      testLogFile.close(isRemoveLogs);
+    } else {
+      logFile.close(true);
+    }
   }
 
   /**
-   * Wrapper for read/write the operation log file
+   * Wrapper for read the operation log file
    */
   private class LogFile {
-    private File file;
+    private final File file;
     private BufferedReader in;
-    private PrintStream out;
     private volatile boolean isRemoved;
 
-    LogFile(File file) throws FileNotFoundException {
+    LogFile(File file) {
       this.file = file;
-      in = new BufferedReader(new InputStreamReader(new FileInputStream(file)));
-      out = new PrintStream(new FileOutputStream(file));
       isRemoved = false;
-    }
-
-    synchronized void write(String msg) {
-      // write log to the file
-      out.print(msg);
     }
 
     synchronized List<String> read(boolean isFetchFirst, long maxRows)
@@ -152,16 +148,19 @@ public class OperationLog {
       return readResults(maxRows);
     }
 
-    void remove() {
+    /**
+     * Close the logs, and remove them if specified.
+     * @param removeLog If true, remove the log file
+     */
+    synchronized void close(boolean removeLog) {
       try {
         if (in != null) {
           in.close();
         }
-        if (out != null) {
-          out.close();
+        if (!isRemoved && removeLog && file.exists()) {
+          FileUtils.forceDelete(file);
+          isRemoved = true;
         }
-        FileUtils.forceDelete(file);
-        isRemoved = true;
       } catch (Exception e) {
         LOG.error("Failed to remove corresponding log file of operation: " + operationName, e);
       }
@@ -169,27 +168,27 @@ public class OperationLog {
 
     private void resetIn() {
       if (in != null) {
-        IOUtils.cleanup(LOG, in);
+        IOUtils.closeStream(in);
         in = null;
       }
     }
 
     private List<String> readResults(long nLines) throws SQLException {
+      List<String> logs = new ArrayList<String>();
       if (in == null) {
         try {
           in = new BufferedReader(new InputStreamReader(new FileInputStream(file)));
-        } catch (FileNotFoundException e) {
-          if (isRemoved) {
-            throw new SQLException("The operation has been closed and its log file " +
-                file.getAbsolutePath() + " has been removed.", e);
-          } else {
-            throw new SQLException("Operation Log file " + file.getAbsolutePath() +
-                " is not found.", e);
+          // Adding name of the log file in an extra log line, so it is easier to find
+          // the original if there is a test error
+          if (isShortLogs) {
+            logs.add("Reading log file: " + file);
+            nLines--;
           }
+        } catch (FileNotFoundException e) {
+          return logs;
         }
       }
 
-      List<String> logs = new ArrayList<String>();
       String line = "";
       // if nLines <= 0, read all lines in log file.
       for (int i = 0; i < nLines || nLines <= 0; i++) {
